CHANGE: Changed how the logs are structured
The log page on the Data Terminal now read from a DataTable, called DT_Logs

In the BP for the data terminal the old log array has been replaced with a new variable, Log Array. To populate the logpage, fill this array with the Row Name(s) from the datatable of the logs you want the terminal to display.