Menu update: Continue button, credits, settings
Made a new page for the credits thats static. Added Futuregames and UE logos to the project and credits. Added Karin to credits.

Cleaned up the settings menu a bit. Its now less tall.

Some visual tweaks to the main menu, renamed the Load game button to Continue. Made it so the button is only visable when a savefile exists.

Updated pause menu to follow main menus new measurements.